+++ include/render/primitives/Disk.h
@@ -26,6 +26,8 @@ namespace render {
 
     virtual const Primitive* intersect(const Rayd& ray, HitPointInterval& hitPoints,
                                        render::State& state) const override;
+    PrimitivePacketHit4 intersectPacketHits(const Ray4& rays,
+                                            const PrimitivePacketState4& states) const override;
     virtual Vector3d farthestPoint(const Vector3d& direction) const override;
 
     /**